+++ services/identity-service/src/bootstrap/broker.ts
@@ -1,9 +1,9 @@
 import { NatsBroker } from "@pine/events";
 import { logger } from "@/bootstrap/logger";
-import { env } from "@/env";
+import { env } from "@/bootstrap/env";
 
 export const broker = new NatsBroker({
-  servers: [env.NATS_CLUSTER_URL || "nats"],
+  servers: [env.NATS_URL || "nats"],
   streams: ["user"],
   logger,
 });