+++ services/identity-service/src/bootstrap/index.ts
@@ -1,7 +1,8 @@
 export { broker } from "@/bootstrap/broker";
 export { container } from "@/bootstrap/container";
+export { TYPES } from "@/bootstrap/container-types";
 export { closeDb, db, initializeDb } from "@/bootstrap/db";
+export { env } from "@/bootstrap/env";
 export { hydraClient } from "@/bootstrap/hydra-client";
 export { kratosClient } from "@/bootstrap/kratos-client";
 export { logger } from "@/bootstrap/logger";
-export { TYPES } from "@/bootstrap/container-types";