+++ services/inventory-service/src/features/me/services/index.ts
@@ -1,0 +1,2 @@
+export * from "@/features/me/services/IMeService";
+export * from "@/features/me/services/MeService";